refactor(friends): decongest friends page by moving code to components
